kgea/server/kgea_archive_requirements.txt
connexion[aiohttp,swagger-ui] >= 2.6.0; python_version>="3.6"
# 2.3 is the last version that supports python 3.5
connexion[aiohttp,swagger-ui] <= 2.3.0; python_version=="3.5" or python_version=="3.4"
# connexion requires werkzeug but connexion < 2.4.0 does not install werkzeug
# we must peg werkzeug versions below to fix connexion
# https://github.com/zalando/connexion/pull/1044
werkzeug == 0.16.1; python_version=="3.5" or python_version=="3.4"
swagger-ui-bundle == 0.0.6
aiohttp_jinja2 == 1.2.0
jinja2 == 2.11.3
aiohttp_cors >= 0.7.0

# KGE specific
botocore<1.21.0,>=1.20.12
boto3 >= 1.17.0
pyyaml
#flask
#Flask-Session
#flask-cors >= 3.0.10
aiohttp
aiomcache
